+++ lib/import.sh
@@ -4,3 +4,4 @@ export __tfm_lib_dir="${__tfm_root_dir}/lib"
 
 # import TF wrapper modules
 source "${__tfm_lib_dir}/config_parse.sh"
+source "${__tfm_lib_dir}/validate.sh"